Update README.md (#208)
* Update README.md

Add backstage-issues.md to README.

Signed-off-by: Dave Welsch <116022979+dwelsch-esi@users.noreply.github.com>

* Add Glossary to the Backstage analysis directory

Add a glossary of terms to the Backstage analysis. Can be used as the basis for a comprehensive glossary in the Backstage technical documentation. Contains terms for the Backstage project, the Backstage product, and the CNCF tech doc analysis, so will need editing first. Also: not necessarily complete, but a good starting point.

Signed-off-by: Dave Welsch <116022979+dwelsch-esi@users.noreply.github.com>

* Add Backstage Glossary to list of docs in README.md

This is the glossary I compiled while doing the tech doc analysis. Can be used as the basis for a glossary in the Backstage documentation, but should be edited first.

Signed-off-by: Dave Welsch <116022979+dwelsch-esi@users.noreply.github.com>

---------

Signed-off-by: Dave Welsch <116022979+dwelsch-esi@users.noreply.github.com>